+++ skbio/io/__init__.py
@@ -23,6 +23,7 @@ are supported by each format, see the associated documentation.
    blast7
    clustal
    embl
+   embed
    fasta
    fastq
    genbank
@@ -245,6 +246,7 @@ from ._exception import (
     GFF3FormatError,
     EMBLFormatError,
     BIOMFormatError,
+    EmbedFormatError,
 )
 from .registry import write, read, sniff, create_format, io_registry
 from .util import open
@@ -276,6 +278,7 @@ __all__ = [
     "QUALFormatError",
     "StockholmFormatError",
     "BIOMFormatError",
+    "EmbedFormatError",
 ]
 
 
@@ -300,6 +303,7 @@ import_module("skbio.io.format.binary_dm")
 import_module("skbio.io.format.taxdump")
 import_module("skbio.io.format.sample_metadata")
 import_module("skbio.io.format.biom")
+import_module("skbio.io.format.embed")
 
 # This is meant to be a handy indicator to the user that they have done
 # something wrong.